seperated network from node lib
